recipes/runner/level-1-basic/simple-workflow/pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "arazzo-recipe-simple-workflow"
version = "0.1.0"
description = "Simple workflow recipe for Arazzo Engine Cookbook"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arazzo-runner<=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
# Don't try to build a wheel - just install dependencies
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v"]